pending now handles the emitter identification on the client
it was responsible for addListener requests only before.

all client requests are now expressed as a pending request, that as
soon as it is validated and confirmed by the server, reflected on the
client.

this is necessary since the client does not actually know which object
is targeted by an event request, so it has to ask the server before an
event operation can be executed locally.

TODO: see previous commit