completely restructured Plugin update & install
install implementation was not really fit to be used for updates.

This is changed now. Instead of a map that is given to install, install
gets a target getter, which evaluates to which Plugin a manifest should
be assigned (if at all). This happens before the installation actually
starts so that the target getter can stop an installation. This was
necessary to accommodate for the various rules that apply to different
ways of installation.