Added Runner#(enabled|enable|disable)
The enabled property specifies whether a runner is activatable or not.

If a runner is disabled it will be deactivated as well. Enabling a
runner will not activate it though.

„enabled“ is true by default and is not writable via owe, since
enabling and disabling is used only by core to specifically prevent
users from activating runners during critical operations. Giving the
user to ability to reenable a disabled runner would render this feature
useless.