Runners cannot be explicitly enabled/disabled anymore
Since runners are only ever temporarily disabled due to a specific
operation being performed, runners are disabled by handing a promise to
Runner#disableAsLongAs, which causes the runner to be disabled as long
as the given promise is not settled (uses a PromiseQueue internally).

This implementation makes to possible to have multiple parallel
operations that all block runner activation. Only if all operations are
completed, the runner will be reenabled.