Runner#activate rejects if disabled (partial revert of prev commit)
The reason rejection was disabled was to prevent the start-stop-start
behavior described in the previous commit.

Delaying activation when disabled is the wrong behavior though because
a runner is usually disabled while breaking changes are applied to it.
It would not be safe to simply delay activation, since the user may not
actually want this to happen once the runner is reenabled.

The start-stop-start bug was fixed by introducing a special case for
initial activations: Those set active to true immediately but do not
start a sandbox yet. Plugin update will take note of such activations
(since only Plugin#active is checked) and will make sure the runner is
actually activated after an update (if it’s safe). If a runner has to
autoupdating plugin dependencies, the initial Plugin#activate call will
then simply activate the runner, after its graph is loaded (all
autoupdates completed).