moved to an async graph assign implementation
Graph#assign is now performed in two stageManager-managed phases: "nodes" and "edges". This causes ALL nodes (the nodes of all graphs) to be assigned before any edge is set.
This new behavior fixes a bug in the previous implementation: When validating nested graphs a race condition would occur if the outer graph was assigned before the inner graph. When checking whether the outer graph's edges are connected valid I/O ports of the inner graph, the checks would always fail, since the inner (not yet assigned) graph was empty and thus looked like a graph without any I/O.

The new two stage design fixes this. The implementation of it changed many areas of the graph implementation. Most notably:
+ Node and all its implementors are Assignables now
+ NodeExecutor and all its implementors are Assignables now